[generator] Files generator/final_processor_intermediate_mwm.* were splitted into several files generator/final_processor_[processor type].*.
